requirements-test.txt
-r requirements.txt  # Ensure tests dependencies are automatically included

# Put dependencies required for testing the module here
pytest-mock ~= 3.14
